corrections for paranoid

+++ src/ArbConnectivity.hpp
@@ -34,12 +34,12 @@ struct ArbLatticeConnectivity {
         FILE* f;
         f = fopen(filename.c_str(),"w");
         fprintf(f,"idx_og,idx");
-        for (int q=0;q<Q;q++) fprintf(f,",nbr%d",q);
+        for (size_t q=0;q<Q;q++) fprintf(f,",nbr%ld",q);
         fprintf(f,"\n");
         size_t n = chunk_end - chunk_begin;
         for (size_t lid=0; lid<n; lid++) {
             fprintf(f,"%ld,%ld",(size_t) og_index[lid],(size_t) lid + chunk_begin);
-            for (int q=0;q<Q;q++) fprintf(f,",%ld",(signed long int) neighbor(q, lid));
+            for (size_t q=0;q<Q;q++) fprintf(f,",%ld",(signed long int) neighbor(q, lid));
             fprintf(f,"\n");
         }
         fclose(f);
